# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Drydock API",
    "version": "1.1.0",
    "description": "Automation surface for Drydock. Authenticate with an organization API key (`Authorization: Bearer ddk_…`); see docs/api-keys.md. Every key reads; a key with `scan` access can also start reviews (operations marked `x-drydock-api-key-access: scan`). No key records a release decision. Package-diff endpoints are anonymous.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "https://drydock.org",
      "description": "Hosted Drydock"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
...
  "paths": {
...
    "/api/v1/scans": {
      "get": {
        "operationId": "listScans",
        "summary": "List the organization's reviews",
        "description": "Newest first. `filter` defaults to `undecided`.",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Decision filter",
            "schema": {
              "type": "string",
              "enum": [
                "undecided",
                "published_without_decision",
                "publish",
                "no_publish",
                "all"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "`nextCursor` from the previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanList"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "startScan",
        "summary": "Start a review",
        "description": "Queues a review of one staged npm publish (`stageId`) or one published release. Needs a key with `scan` access; poll `GET /api/v1/scans/{id}/status` for the result. Starting a review never records a release decision.",
        "tags": [
          "Reviews"
        ],
        "x-drydock-api-key-access": "scan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartScanRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartedScan"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "The review has not completed, or the package is claimed by another organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "The registry named a package Drydock cannot review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; honor `retry-after`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "The registry rejected or could not answer the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "The registry could not be verified; retry later",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staged-publishes/scan": {
      "post": {
        "operationId": "checkNpmForStagedPublishes",
        "summary": "Check npm for staged publishes",
        "description": "The dashboard's \"Check npm\": lists the organization's staged npm publishes with its npm token and starts a review of each one not reviewed yet. Needs a key with `scan` access.",
        "tags": [
          "Reviews"
        ],
        "x-drydock-api-key-access": "scan",
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StagedPublishDiscovery"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; honor `retry-after`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "The registry rejected or could not answer the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/scans/{id}": {
      "get": {
        "operationId": "getScan",
        "summary": "Read one review",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Review (scan) id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDetail"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/scans/{id}/status": {
      "get": {
        "operationId": "getScanStatus",
        "summary": "A review's lifecycle status, for polling",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Review (scan) id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanStatus"
                }
              }
            }
          },
          "401": {
            "description": "Missing, unknown, or expired credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not allowed for this credential or organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found in this organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ddk_…",
        "description": "Organization API key from Settings → Integrations → API keys"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "description": "Stable reason a client can act on",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": {}
      },
...
      "ScanList": {
        "type": "object",
        "properties": {
          "scans": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "stageId": {
                  "type": "string"
                },
                "source": {
                  "type": "string",
                  "enum": [
                    "manual",
                    "auto_discovery",
                    "workflow_gate",
                    "published"
                  ]
                },
                "packageName": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "stagedVersion": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "previousVersion": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "risk": {
                  "type": "string",
                  "description": "`low`, `medium`, `high`, or `critical`; `pending` until a review completes"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "pending",
                    "running",
                    "complete",
                    "failed"
                  ]
                },
                "decision": {
                  "anyOf": [
                    {
                      "type": "string",
                      "enum": [
                        "publish",
                        "no_publish"
                      ]
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "decisionReason": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "decidedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "description": "ISO 8601 timestamp"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "changedFileCount": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "findingCount": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "reportDigest": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "startedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "description": "ISO 8601 timestamp"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "completedAt": {
                  "anyOf": [
                    {
                      "type": "string",
                      "description": "ISO 8601 timestamp"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "createdAt": {
                  "type": "string",
                  "description": "ISO 8601 timestamp"
                },
                "updatedAt": {
                  "type": "string",
                  "description": "ISO 8601 timestamp"
                },
                "ecosystem": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "riskSummary": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "artifactRisk": {
                          "type": "string"
                        },
                        "releaseRisk": {
                          "type": "string"
                        },
                        "contextRisk": {
                          "type": "string"
                        },
                        "releaseFindingCount": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "contextFindingCount": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "unknownFindingCount": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "artifactRisk",
                        "releaseRisk",
                        "contextRisk",
                        "releaseFindingCount",
                        "contextFindingCount",
                        "unknownFindingCount"
                      ],
                      "additionalProperties": {},
                      "description": "Risk split into the release delta and pre-existing package context"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "stageId",
                "source",
                "packageName",
                "stagedVersion",
                "previousVersion",
                "risk",
                "status",
                "decision",
                "decisionReason",
                "decidedAt",
                "changedFileCount",
                "findingCount",
                "reportDigest",
                "startedAt",
                "completedAt",
                "createdAt",
                "updatedAt",
                "ecosystem",
                "riskSummary"
              ],
              "additionalProperties": {}
            }
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `cursor` for the next page"
          },
          "filter": {
            "type": "string",
            "enum": [
              "undecided",
              "published_without_decision",
              "publish",
              "no_publish",
              "all"
            ]
          },
          "limit": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "scans",
          "nextCursor",
          "filter",
          "limit"
        ],
        "additionalProperties": {}
      },
...
    }
  }
}