# server/routes/scans/lifecycle.ts
/**
 * Scan lifecycle: start one, list them, read one back, delete a failed one.
 *
 * Everything here is organization-scoped; a scan is only ever visible to the
 * organization that created it. `POST /` is rate limited per organization
 * because starting a scan spends registry egress and queue budget.
 */
import { Hono } from "hono";
import {
  ORGANIZATION_SCAN_LIMIT,
  ORGANIZATION_SCAN_WINDOW_MS,
  guardRateLimit,
} from "../../lib/rate-limit";
import type { AppDb } from "../../db/client";
import { getNpmConnection } from "../../db/npm-connections";
import { recordScanEvent } from "../../db/events";
import {
  LIST_SCANS_DEFAULT_LIMIT,
  LIST_SCANS_MAX_LIMIT,
  SCAN_DECISION_FILTERS,
  type ScanDecisionFilter,
  type ScanSource,
  createScanJob,
  deleteFailedScan,
  getScan,
  getScanFile,
  getScanStatus,
  listScans,
} from "../../db/scans";
import { recordApiKeyAction } from "../../db/api-keys";
import { requestActorUserId, requireActiveOrganization } from "../../lib/auth/active-organization";
import { deleteScanArtifacts, scanArtifactReadBucket } from "../../lib/scan/artifacts";
import { canonicalOrigin, parseLimitQuery, readJsonObject } from "../../lib/platform/http";
import { workerExecutionContext } from "../../lib/platform/execution-context";
import { allowInsecureLocalRegistry, decryptNpmToken } from "../../lib/ecosystems/npm/connection";
import {
  checkStagedPublishAccess,
  fetchStagedPublishDetails,
} from "../../lib/ecosystems/npm/staged-publishes";
import { getPublicationMonitor, getPublishedAdapter, getStagedAdapter } from "../../lib/ecosystems";
import { publishedPairStageId } from "../../lib/ecosystems/published-pair";
import { PublicDiffError } from "../../lib/public-diff/error";
import { parseScanInput, type PublishedScanRequest } from "../../lib/scan/input";
import { executeScanJob, type ScanQueueMessage } from "../../lib/scan/job";
import { encodeListScansCursor, parseListScansCursor } from "../../lib/scan/list-cursor";
import { recordProductEvent } from "../../lib/analytics";
import { describeOperationalError, emitOperationalEvent } from "../../lib/platform/observability";
import type { Bindings, ScanInput, Variables } from "../../types";
import { PackageClaimConflictError } from "../../db/package-claims";
import { readNpmPackageClaimAvailability } from "../../db/scan-jobs";

export const scanLifecycleRoutes = new Hono<{ Bindings: Bindings; Variables: Variables }>();

scanLifecycleRoutes.post("/", async (c) => {
  const body = await readJsonObject<ScanInput>(c);
  const parsed = parseScanInput(body);
  if (!parsed.ok) return c.json({ error: parsed.error }, parsed.status);

  const db = c.var.db;
  // A `scan` API key acts as its creator (`requestActorUserId`).
  const actorUserId = requestActorUserId(c);
  const organizationId = await requireActiveOrganization(c, db);
  const limited = await guardRateLimit(
    c,
    {
      key: `scan:${organizationId}`,
      limit: ORGANIZATION_SCAN_LIMIT,
      windowMs: ORGANIZATION_SCAN_WINDOW_MS,
    },
    "scan rate limit exceeded",
  );
  if (limited) return limited;

  const prepared =
    parsed.kind === "published"
      ? await preparePublishedScan(c, parsed.request)
      : await prepareStagedScan(c, db, organizationId, parsed.input);
  if ("error" in prepared) return prepared.error;

  const scanId = crypto.randomUUID();
  const detail = await createScanJob(db, {
    id: scanId,
    stageId: prepared.input.stageId,
    organizationId,
    ownerUserId: actorUserId,
    source: prepared.source,
    packageName: prepared.packageName,
    stagedVersion: prepared.version,
    stagedCreatedAt: prepared.stagedCreatedAt,
    stagedDeclaredSha1: prepared.stagedDeclaredSha1,
    registryUrl: prepared.registryUrl,
    stageAccessStatus: prepared.stageAccessStatus,
  }).catch((err: unknown) => {
    if (err instanceof PackageClaimConflictError) return err;
    throw err;
  });
  if (detail instanceof PackageClaimConflictError) {
    // The caller already proved stage access; only its own history is described.
    const availability = await readNpmPackageClaimAvailability(db, {
      registryUrl: prepared.registryUrl!,
      packageName: prepared.packageName!,
      organizationId,
    });
    const error =
      availability === "own_history"
        ? "This organization's earlier reviews of this package are awaiting an ownership confirmation by support. New reviews can start once support confirms it."
        : detail.message;
    return c.json({ error }, 409);
  }
  if (!detail) return c.json({ error: "failed to create scan" }, 500);
  if (prepared.staged) {
    await getPublicationMonitor("npm")?.registerStagedReleases(db, c.env, {
      organizationId,
      registryUrl: prepared.registryUrl!,
      releases: [prepared.staged],
    });
  }
  const message: ScanQueueMessage = {
    ...prepared.input,
    scanId,
    organizationId,
    actorUserId,
    source: prepared.source,
  };

  // Counted at creation, not completion, so the queued → completed drop-off
  // is visible: a scan that never reaches a terminal state emits neither
  // `scan.completed` nor `scan.failed` and would otherwise vanish.
  recordProductEvent(c.env, {
    name: "scan.queued",
    organizationId,
    ecosystem: prepared.ecosystem,
    source: message.source ?? "manual",
  });

  if (c.env.SCAN_QUEUE) {
    await c.env.SCAN_QUEUE.send(message);
  } else {
    c.executionCtx.waitUntil(
      executeScanJob(c.env, workerExecutionContext(c.executionCtx), message, db, {
        finalAttempt: true,
      }),
    );
  }

  const apiKey = c.get("apiKey");
  if (apiKey) {
    await recordApiKeyAction(db, apiKey, {
      type: "organization.api_key_review_started",
      scanId,
      metadata: {
        ecosystem: prepared.ecosystem,
        packageName: prepared.packageName,
        stagedVersion: prepared.version,
      },
    });
  }

  return c.json({ scan: detail?.scan, queued: Boolean(c.env.SCAN_QUEUE) }, 202);
});

type ScanRouteContext = import("hono").Context<{ Bindings: Bindings; Variables: Variables }>;

interface PreparedScan {
  input: ScanInput;
  source: ScanSource;
  ecosystem: string;
  packageName: string | null;
  version: string | null;
  /**
   * Registry-reported stage creation time. Only a staged npm scan has one; a
   * published-pair review was never staged.
   */
  stagedCreatedAt: string | null;
  /** The registry's SHA-1 from the same stage record; only a staged npm scan has one. */
  stagedDeclaredSha1: string | null;
  /**
   * Only a staged npm scan captures one. A published-pair review must leave it
   * null: `createScanJob` uses it to claim the registry coordinates a staged
   * release owns, and a review of an already-public version has no claim on them.
   */
  registryUrl: string | null;
  stageAccessStatus?: number | null;
  staged?: NonNullable<Awaited<ReturnType<typeof fetchStagedPublishDetails>>>;
}

async function prepareStagedScan(
  c: ScanRouteContext,
  db: AppDb,
  organizationId: string,
  input: ScanInput,
): Promise<PreparedScan | { error: Response }> {
  const npmConnection = await getNpmConnection(db, organizationId);
  if (!npmConnection) {
    return {
      error: c.json(
        { error: "Connect an organization npm token before scanning staged publishes." },
        400,
      ),
    };
  }
  if (npmConnection.validationStatus !== "valid") {
    return {
      error: c.json(
        { error: "Validate the organization npm token before scanning staged publishes." },
        400,
      ),
    };
  }
  const token = await decryptNpmToken(c.env, npmConnection);
  const access = await checkStagedPublishAccess(npmConnection.registryUrl, token, input.stageId, {
    allowInsecureLocalhost: allowInsecureLocalRegistry(c.env),
  });
  if (!access.allowed) {
    return {
      error: c.json(
        {
          error: "This organization's npm token cannot access that staged publish.",
          status: access.status,
        },
        403,
      ),
    };
  }

  if (access.status === null || access.status < 200 || access.status >= 300) {
    return { error: c.json({ error: "Could not verify npm stage access. Try again later." }, 503) };
  }

  // Package claims need a registry identity before admission; manifest bytes
  // cannot grant ownership and a transient metadata failure must not reserve it.
  const staged = await fetchStagedPublishDetails(npmConnection.registryUrl, token, input.stageId, {
    allowInsecureLocalhost: allowInsecureLocalRegistry(c.env),
  }).catch(() => null);

  if (!staged?.packageName || staged.id !== input.stageId || !staged.version?.trim()) {
    return {
      error: c.json({ error: "Could not verify npm package identity. Try again later." }, 503),
    };
  }
  // npm answered with a complete stage record; a name that still fails
  // validation is permanent, so retrying would never help.
  if (
    !getStagedAdapter("npm").stagedClaimIdentity?.({
      registryUrl: npmConnection.registryUrl,
      packageName: staged.packageName,
      version: staged.version,
    })
  ) {
    return {
      error: c.json(
        { error: "npm reported a package name for this stage that Drydock cannot review." },
        422,
      ),
    };
  }

  return {
    input,
    source: "manual",
    ecosystem: "npm",
    packageName: staged?.packageName ?? null,
    version: staged?.version ?? null,
    stagedCreatedAt: staged?.createdAt ?? null,
    stagedDeclaredSha1: staged?.shasum ?? null,
    registryUrl: npmConnection.registryUrl,
    stageAccessStatus: access.status,
    staged,
  };
}

/**
 * Resolve a published `package@version` against its public registry before any
 * scan row exists, so an unpublished version is a request error rather than a
 * scan that fails minutes later, and the queued message names an exact pair.
 *
 * No npm credential is read, decrypted, or attached here: acquisition reuses
 * the same public brokers the anonymous `/diff` surface uses.
 */
async function preparePublishedScan(
  c: ScanRouteContext,
  request: PublishedScanRequest,
): Promise<PreparedScan | { error: Response }> {
  const adapter = getPublishedAdapter(request.ecosystem);
  if (!adapter) return { error: c.json({ error: "unsupported ecosystem" }, 400) };

  let resolved: Awaited<ReturnType<typeof adapter.resolvePair>>;
  try {
    resolved = await adapter.resolvePair(c.env, workerExecutionContext(c.executionCtx), request);
  } catch (err) {
    emitOperationalEvent("warn", "scan.published_pair.resolve_failed", {
      ecosystem: request.ecosystem,
      error: describeOperationalError(err),
    });
    // The public-diff loaders already classify their own failures (unknown
    // package, oversized archive, registry unreachable) with a public-safe
    // message and status; anything else is ours and stays opaque.
    if (err instanceof PublicDiffError) {
      return { error: c.json({ error: err.message }, err.status) };
    }
    return { error: c.json({ error: "could not reach the registry for that package" }, 502) };
  }
  if (!resolved.ok) return { error: c.json({ error: resolved.error }, resolved.status) };

  const pair = resolved.pair;
  return {
    input: { stageId: publishedPairStageId(pair), published: pair },
    source: "published",
    ecosystem: pair.ecosystem,
    packageName: pair.packageName,
    version: pair.version,
    stagedCreatedAt: null,
    stagedDeclaredSha1: null,
    registryUrl: null,
  };
}

const DECISION_FILTER_SET = new Set<ScanDecisionFilter>(SCAN_DECISION_FILTERS);

scanLifecycleRoutes.get("/", async (c) => {
  const db = c.var.db;
  const organizationId = await requireActiveOrganization(c, db);

  const rawFilter = c.req.query("filter");
  const decisionFilter: ScanDecisionFilter = DECISION_FILTER_SET.has(
    rawFilter as ScanDecisionFilter,
  )
    ? (rawFilter as ScanDecisionFilter)
    : "undecided";

  const limit = parseLimitQuery(c.req.query("limit"), {
    default: LIST_SCANS_DEFAULT_LIMIT,
    max: LIST_SCANS_MAX_LIMIT,
  });

  const cursor = parseListScansCursor(c.req.query("cursor"));

  const result = await listScans(db, organizationId, { cursor, limit, decisionFilter });
  return c.json({
    scans: result.scans,
    nextCursor: encodeListScansCursor(result.nextCursor),
    filter: decisionFilter,
    limit,
  });
});

scanLifecycleRoutes.delete("/:id", async (c) => {
  const db = c.var.db;
  const session = c.get("authSession");
  const organizationId = await requireActiveOrganization(c, db);
  const scanId = c.req.param("id");

  const result = await deleteFailedScan(db, scanId, organizationId);
  if (result.outcome === "not_found") return c.json({ error: "not found" }, 404);
  if (result.outcome === "not_failed") {
    return c.json({ error: "only failed scans can be deleted" }, 409);
  }

  await Promise.all([
    deleteScanArtifacts(c.env.ARTIFACTS, organizationId, scanId),
    recordScanEvent(db, {
      organizationId,
      actorUserId: session.userId,
      type: "scan.deleted",
      metadata: { scanId, status: "failed", source: result.source },
    }),
  ]);
  return c.json({ ok: true, id: scanId });
});

// Public share link management. Enabling exposes the completed scan's
// canonical report export (and its signed attestation) at
// /public/reports/:token to anyone holding the link — an explicit, elevated

scanLifecycleRoutes.get("/:id", async (c) => {
  const db = c.var.db;
  const organizationId = await requireActiveOrganization(c, db);
  const scan = await getScan(db, c.req.param("id"), organizationId, scanArtifactReadBucket(c.env), {
    files: "list",
  });
  if (!scan) return c.json({ error: "not found" }, 404);
  return c.json({
    ...scan,
    scan: {
      ...scan.scan,
      publicShareUrl: scan.scan.publicShareToken
        ? `${canonicalOrigin(c)}/reports/${scan.scan.publicShareToken}`
        : null,
      publicShareIncludesFiles: scan.scan.publicShareIncludesFiles,
    },
  });
});

scanLifecycleRoutes.get("/:id/status", async (c) => {
  const db = c.var.db;
  const organizationId = await requireActiveOrganization(c, db);
  const scan = await getScanStatus(db, c.req.param("id"), organizationId);
  if (!scan) return c.json({ error: "not found" }, 404);
  return c.json({ scan });
});

scanLifecycleRoutes.get("/:id/file", async (c) => {
  const path = c.req.query("path") || "";
  if (!path) return c.json({ error: "path is required" }, 400);
  const db = c.var.db;
  const organizationId = await requireActiveOrganization(c, db);
  const file = await getScanFile(
    db,
    c.req.param("id"),
    organizationId,
    path,
    scanArtifactReadBucket(c.env),
  );
  if (!file) return c.json({ error: "file not found in scan" }, 404);
  return c.json({ file }, 200, { "cache-control": "private, max-age=300" });
});